# perf: faster `LocalProxy`