#### before

```python
In [4]: %timeit frappe.db.set_value
919 ns ± 6.2 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

In [6]: %timeit frappe.db.set_value
929 ns ± 1.63 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```

#### after (~350ns / 40% reduction in getattr)

```python
In [1]: %timeit frappe.db.set_value
587 ns ± 1.11 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

In [3]: %timeit frappe.db.set_value
585 ns ± 1.52 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

```

#### for context

```python
In [1]: %timeit frappe.local.db.set_value
399 ns ± 1.81 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```