@ankush 2025-01-20
~1.24x faster for me 

before : utils_local_proxy: Mean +- std dev: 434 ns +- 7 ns
after  : utils_local_proxy: Mean +- std dev: 350 ns +- 8 ns